operator: Refactoring the directory structure
By default operator-sdk treats the generated operator as a dedicated
repository/directory. But in our case we keep operator code as part of main
pmem-csi source repo, instead of dedicated repo/directory.

We do below changes:
  - operator/pkg/apis -> pmem-csi/pkg/apis
  - operator/pkg/controller -> pmem-csi/pkg/pmem-csi-operator/controller
  - operator/manager/main.go -> pkg/pmem-csi-operator/main.go

In addition to that we had to manage below links to satisfy operator-sdk code
generator tool:
- cmd/manager -> cmd/pmem-csi-operator (expects cmd/manger/main.go)
- build/Dockerfile -> Dockerfile (expects build/Dockerfile)

Updated vendor tree and runtime dependency trarges with new package/module
dependencies.
