+++ cmd/manager
@@ -1,0 +1,1 @@
+./pmem-csi-operator/
