+++ cmd/pmem-csi-operator/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"flag"
+	"os"
+
+	"github.com/intel/pmem-csi/pkg/pmem-csi-operator"
+)
+
+func main() {
+	flag.Parse()
+	os.Exit(pmemoperator.Main())
+}
