+++ operator/examples/pmem-csi.intel.com_v1alpha1_deployment_cr.yaml
