--- operator/tools.go
@@ -1,5 +1,0 @@
-// +build tools
-
-// Place any runtime dependencies as imports in this file.
-// Go modules will be forced to download and install them.
-package tools
