--- operator/version/version.go
@@ -1,5 +1,0 @@
-package version
-
-var (
-	Version = "0.0.1"
-)
