+++ pkg/apis/addtoscheme_pmemcsi_v1alpha1.go
@@ -1,7 +1,7 @@
 package apis
 
 import (
-	"github.com/intel/pmem-csi/operator/pkg/apis/pmemcsi/v1alpha1"
+	"github.com/intel/pmem-csi/pkg/apis/pmemcsi/v1alpha1"
 )
 
 func init() {
